refactor: move utils to another file
