+++ src/cli/index.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const chalk = require("chalk");
-const dashify = require("dashify");
 const fs = require("fs");
 const getStream = require("get-stream");
 const globby = require("globby");
@@ -14,6 +13,7 @@ const prettier = eval("require")("../../index");
 const cleanAST = require("../clean-ast").cleanAST;
 const resolver = require("../resolve-config");
 const constant = require("./constant");
+const util = require("./util");
 
 function run(args) {
   const argv = minimist(args, constant.options);
@@ -57,103 +57,14 @@ function run(args) {
             "bracket-spacing": true,
             parser: "babylon"
           },
-          dashifyObject(options)
+          util.dashifyObject(options)
         )
       });
-      return getOptions(Object.assign({}, argv, parsedArgs));
+      return util.getOptions(Object.assign({}, argv, parsedArgs));
     } catch (error) {
       console.error("Invalid configuration file:", error.toString());
       process.exit(2);
     }
-  }
-
-  function getOptions(argv) {
-    return {
-      cursorOffset: getIntOption(argv, "cursor-offset"),
-      rangeStart: getIntOption(argv, "range-start"),
-      rangeEnd: getIntOption(argv, "range-end"),
-      useTabs: argv["use-tabs"],
-      semi: argv["semi"],
-      printWidth: getIntOption(argv, "print-width"),
-      tabWidth: getIntOption(argv, "tab-width"),
-      bracketSpacing: argv["bracket-spacing"],
-      singleQuote: argv["single-quote"],
-      jsxBracketSameLine: argv["jsx-bracket-same-line"],
-      filepath: argv["stdin-filepath"],
-      trailingComma: getTrailingComma(argv),
-      parser: getParserOption(argv)
-    };
-  }
-
-  function getParserOption(argv) {
-    const value = argv.parser;
-
-    if (value === undefined) {
-      return value;
-    }
-
-    // For backward compatibility. Deprecated in 0.0.10
-    if (argv["flow-parser"]) {
-      console.warn(
-        "`--flow-parser` is deprecated. Use `--parser flow` instead."
-      );
-      return "flow";
-    }
-
-    return value;
-  }
-
-  function getIntOption(argv, optionName) {
-    const value = argv[optionName];
-
-    if (value === undefined) {
-      return value;
-    }
-
-    if (/^\d+$/.test(value)) {
-      return Number(value);
-    }
-
-    console.error(
-      "Invalid --" +
-        optionName +
-        " value. Expected an integer, but received: " +
-        JSON.stringify(value)
-    );
-    process.exit(1);
-  }
-
-  function getTrailingComma(argv) {
-    switch (argv["trailing-comma"]) {
-      case undefined:
-      case "none":
-        return "none";
-      case "":
-        console.warn(
-          "Warning: `--trailing-comma` was used without an argument. This is deprecated. " +
-            'Specify "none", "es5", or "all".'
-        );
-        return "es5";
-      case "es5":
-        return "es5";
-      case "all":
-        return "all";
-      default:
-        throw new Error("Invalid option for --trailing-comma");
-    }
-  }
-
-  function dashifyObject(object) {
-    return Object.keys(object || {}).reduce((output, key) => {
-      output[dashify(key)] = object[key];
-      return output;
-    }, {});
-  }
-
-  function diff(a, b) {
-    return require("diff").createTwoFilesPatch("", "", a, b, "", "", {
-      context: 2
-    });
   }
 
   function format(input, opt) {
@@ -167,7 +78,7 @@ function run(args) {
       const pppp = prettier.format(pp, opt);
       if (pp !== pppp) {
         throw "prettier(input) !== prettier(prettier(input))\n" +
-          diff(pp, pppp);
+          util.diff(pp, pppp);
       } else {
         const ast = cleanAST(prettier.__debug.parse(input, opt));
         const past = cleanAST(prettier.__debug.parse(pp, opt));
@@ -177,41 +88,17 @@ function run(args) {
           const astDiff =
             ast.length > MAX_AST_SIZE || past.length > MAX_AST_SIZE
               ? "AST diff too large to render"
-              : diff(ast, past);
+              : util.diff(ast, past);
           throw "ast(input) !== ast(prettier(input))\n" +
             astDiff +
             "\n" +
-            diff(input, pp);
+            util.diff(input, pp);
         }
       }
       return { formatted: opt.filepath || "(stdin)\n" };
     }
 
     return prettier.formatWithCursor(input, opt);
-  }
-
-  function handleError(filename, e) {
-    const isParseError = Boolean(e && e.loc);
-    const isValidationError = /Validation Error/.test(e && e.message);
-
-    // For parse errors and validation errors, we only want to show the error
-    // message formatted in a nice way. `String(e)` takes care of that. Other
-    // (unexpected) errors are passed as-is as a separate argument to
-    // `console.error`. That includes the stack trace (if any), and shows a nice
-    // `util.inspect` of throws things that aren't `Error` objects. (The Flow
-    // parser has mistakenly thrown arrays sometimes.)
-    if (isParseError) {
-      console.error(filename + ": " + String(e));
-    } else if (isValidationError) {
-      console.error(String(e));
-      // If validation fails for one file, it will fail for all of them.
-      process.exit(1);
-    } else {
-      console.error(filename + ":", e.stack || e);
-    }
-
-    // Don't exit the process if one file failed
-    process.exitCode = 2;
   }
 
   if (
@@ -223,7 +110,7 @@ function run(args) {
   }
 
   if (argv["find-config-path"]) {
-    resolveConfig(argv["find-config-path"]);
+    util.resolveConfig(argv["find-config-path"]);
   } else if (stdin) {
     getStream(process.stdin).then(input => {
       const options = getOptionsForFile(process.cwd());
@@ -233,9 +120,9 @@ function run(args) {
       }
 
       try {
-        writeOutput(format(input, options), options);
+        util.writeOutput(format(input, options), options);
       } catch (e) {
-        handleError("stdin", e);
+        util.handleError("stdin", e);
       }
     });
   } else {
@@ -275,7 +162,7 @@ function run(args) {
         // Add newline to split errors from filename line.
         process.stdout.write("\n");
 
-        handleError(filename, e);
+        util.handleError(filename, e);
         return;
       }
 
@@ -312,7 +199,7 @@ function run(args) {
           process.exitCode = 2;
         }
       } else if (!argv["list-different"]) {
-        writeOutput(result, options);
+        util.writeOutput(result, options);
       }
     });
   }
@@ -332,24 +219,6 @@ function run(args) {
     }
 
     return true;
-  }
-
-  function resolveConfig(filePath) {
-    const configFile = resolver.resolveConfigFile.sync(filePath);
-    if (configFile) {
-      console.log(path.relative(process.cwd(), configFile));
-    } else {
-      process.exitCode = 1;
-    }
-  }
-
-  function writeOutput(result, options) {
-    // Don't use `console.log` here since it adds an extra newline at the end.
-    process.stdout.write(result.formatted);
-
-    if (options.cursorOffset) {
-      process.stderr.write(result.cursorOffset + "\n");
-    }
   }
 
   function eachFilename(patterns, callback) {
